deadline_test: Calculate prime numbers in loop and other changes
Instead of doing the getcpu() system call, calculate prime numbers in the
loop, and report the result at the end.

Also, do not exit if test run at setup does not make the deadline. Simply
add the amount that was missed to the overhead, and try again.

Also, fix up the calculation of the test loops when calculating loop time.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>