deadline_test: Update calibration of loops per 1000us
First, remove the code in the run_loops that test the time for a given
input. Because that test just skews the numbers when not doing the test.
Instead, simply start the calibration with 100,000 loops.

Then use that calculation to try to figure out the number of loops needed to
run 1000us. And do that twice.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>