deadline_test: Added option for "step" and updated output
Although the "step" code was already there (the ability to set the step up
in deadline time in microseconds), there was not hook of setting it via the
command line options. Add one.

Also modified the output of the percent, to note that it is per task and not
total.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>